refactor criteria
